tenant config, docker build
tenant config and generator is now its own project in the monorepo. docker build now does not build the app itself, it will be built when starting the container. a tenant config volume needs to be mapped into the container - no config or invalid config will cause the container to fail starting.