PWA: tenant config, docker build, QA workflow, and minor fixes (#135)
* tenant config, docker build

tenant config and generator is now its own project in the monorepo. docker build now does not build the app itself, it will be built when starting the container. a tenant config volume needs to be mapped into the container - no config or invalid config will cause the container to fail starting.

* Adjusted QA workflow for PWA

* use node:24/slim instead of alpine, use npm to add quasar to global packages

* remove icongenie, bump patch and minor versions

* minor fixes and update READMEs

---------

Co-authored-by: Oliver Studer <oliver.studer@tegonal.com>